Optimiser

Only one force optimizer has been implemented, FIRE:

DAC.FIRE — Type

FIRE

Implementation of the fast inertial relaxation engine for energy optimisation.

Arguments

dt::Float64: Some parameter for FIRE. Default values from ASE used.

maxstep::Float64: Some parameter for FIRE. Default values from ASE used.

dtmax::Float64: Some parameter for FIRE. Default values from ASE used.

NMin::Int64: Some parameter for FIRE. Default values from ASE used.

finc::Float64: Some parameter for FIRE. Default values from ASE used.

fdec::Float64: Some parameter for FIRE. Default values from ASE used.

astart::Float64: Some parameter for FIRE. Default values from ASE used.

fa::Float64: Some parameter for FIRE. Default values from ASE used.

a::Float64: Some parameter for FIRE. Default values from ASE used.

downhillCheck::Bool: Some parameter for FIRE. Default values from ASE used.

v::Matrix{Float64}: Some parameter for FIRE. Default values from ASE used.

f2::Matrix{Float64}: Some parameter for FIRE. Default values from ASE used.

f2sum::Vector{Float64}: Some parameter for FIRE. Default values from ASE used.

dr::Matrix{Float64}: Some parameter for FIRE. Default values from ASE used.

DAC.FIRE — Method

FIRE()

Wrapper function, sets all fields of FIRE automatically.

Post Optimisation Tasks

One field for the BasinHopper is postOptimisationTasks, a function called after each optimisation to clean up the Cluster. There are three implemented functions, shown below with their intended use cases explained. One of these functions MUST be passed to the BasinHopper to use the BHA or DACA.

DAC.standardPostOptimisationTasks! — Method

standardPostOptimisationTasks!(cluster::Cluster, bh::BasinHopper)

Sets the CNA profile of a cluster after optimisation. Intended for use with BHA.

DAC.extendedPostOptimisationTasks! — Method

extendedPostOptimisationTasks!(cluster::Cluster, bh::BasinHopper)

Sets the CNA profile of a nanoparticle after optimisation, and does all machine learning work for the DACA so the nanoparticle can be classified. Intended for DACA use when PCA is used in the ML process.

DAC.extendedPostOptimisationTasksNoPCA! — Method

extendedPostOptimisationTasksNoPCA!(cluster::Cluster, bh::BasinHopper)

Sets the CNA profile of a nanoparticle after optimisation, and does all machine learning work for the DACA so the nanoparticle can be classified. Intended for DACA use when PCA is NOT used in the ML process.