DAC.jl Documentation

This is some sample text. Lorem ipsum etc...

DAC.BasinHopper — Type

BasinHopper Struct(optimizer::Optimizer
	calculator::Calculator
	metC::MetC
	reseeder::Reseeder
	formula::Dict{String, Int64}
	boxLength::Float64
	vacuumAdd::Float64
	kT::Float64
	perturber::Function
	postOptimisationTasks::Function
	fmax::Float64
	fmaxTight::Float64
	tightEnergyThreshold::Float64
	rcut::Float64
	energyRounding::Int64
	walltime::Float64
	recordingMode::String
	io::Tuple{IO, Channel}
	logIO::Tuple{IO, Channel}
	CNAIO::Tuple{IO, Channel}
	clusterVector::Union{ClusterVector, ClusterVectorWithML}
	logResumeFile::Bool
	exitOnReseed::Bool
	version::String)

The BasinHopper object underpinning the DACA.

Fields

optimizer::Optimizer: Force optimizer. Use FIRE() or implement your own.

calculator::Calculator: Energy and force calculator. See LJ and RGL.

reseeder::Reseeder: Reseed operator to determine when a reseed is triggered. See Reseeding.

formula::Dict{String, Int64}: Chemical formula. e.g. Dict("Au" => 55).

boxLength::Float64: Box length for generateRandomSeed.

vacuumAdd::Float64: Vacuum Add for generateRandomSeed.

kT::Float64: Redundant (oops).

perturber::Function: Perturber to deform cluster at each hop. See Building Composite Operators

postOptimisationTasks::Function: See Post Optimisation Tasks.

fmax::Float64: Convergence criterion for the force optimiser; maximum force any atom can experience for the optimisation to be considered converged.

fmaxTight::Float64: Optional tighter convergence criterion used if energy found after optimisation converged with fmax is below tightEnergyThreshold. fmaxTight = 0.001 used for Au$_{55}$ to ensure all 16 LESs are distinguishable by their energy.

tightEnergyThreshold::Float64: See above tightEnergyThreshold = -195.18 for Au$_{55}$. If you want this disabled, set to -Inf instead.

rcut::Float64: Used for bond definition for CNA calculation.

energyRounding::Int64: New structures are only stored in clusterVector if no other structure already in the database has the same CNA profile and energy, rounded to the decimal place given by energyRounding.

walltime::Float64: Walltime in hours after which the algorithm will exit. Give Slurm/the given workload manager on the HPC an additional few minutes than given here to ensure a clean and safe exit on timeout.

recordingMode::String: Redundant (oops).

io::Tuple{IO, Channel}: See <ref?>.

logIO::Tuple{IO, Channel}: See <ref?>.

CNAIO::Tuple{IO, Channel}: See <ref?>.

clusterVector::Union{ClusterVector, ClusterVectorWithML}: Database for storing structures.

logResumeFile::Bool: Save a file for resuming the run (probably doesnt work anymore).

exitOnReseed::Bool: true for DACA, false for BHA. DACA handles reseeds in the run.jl file.

version::String: Version check to ensure the user is aware which version of the algorithm is being run