Cluster

The Cluster type

Cluster is a struct (similar to an Object in Python) that stores information about a nanoparticle.

DAC.Cluster — Type

Cluster

Cluster is a type (specifically a mutable struct) that holds information about a nanoparticle. The key fields are shown below (which may be accessed as cluster.field).

Fields

formula: The chemical formula of the cluster, stored as an element keyed Dict type.

positions: The Euclidian coordinates of all atoms in the cluster.

cell: The dimensions of the lattice cell.

energy: The energy of the cluster

energies: The energies of each atom in the cluster.

forces: The forces of each atom in the cluster.

stresses: The stresses of each atom in the cluster.

distances: The distances between each atom in the cluster.

CNA: The total CNA profile of the cluster. No information on the rcut value is stored.

nCNA: The normal CNA profile of the cluster. No information on the rcut value is stored.

atomClassCount: A frequency count of each atom class (e.g. from Atom-64-Class) present in the cluster.

calculator: The calculator used to calculate energies, forces, etc.

Creating Cluster instnaces

Typically, it is impractical to create an instance of Cluster directly, so instead there are various methods capable of creating it. These are described below.

DAC.Cluster — Method

Cluster(formula::Dict{String, Int64})

Creates an almost totally empty instance of Cluster, that only has the chemical formula defined.

Arguments

formula::Dict{String, Int64}: Formula, e.g. Dict("Au" => 55).

DAC.Cluster — Method

Cluster(formula::Dict{String, Int64}, positions::Matrix{Float64})

Creates an instance of Cluster with only the formula and atomic positions defined.

Arguments

formula::Dict{String, Int64}: Formula, e.g. Dict("Au" => 55).

positions::Matrix{Float64}: Atomic positions. As there is no cell, these positions are absolute coordinates.

DAC.Cluster — Method

Cluster(formula::Dict{String, Int64}, positions::Matrix{Float64}, cell::Matrix{Float64})

Creates an instance of Cluster with the formula, atomic positions, and cell dimensions defined

Arguments

formula::Dict{String, Int64}: Formula, e.g. Dict("Au" => 55).

positions::Matrix{Float64}: Atomic positions. As there is no cell, these positions are absolute coordinates.

cell::Matrix{Float64}: The cell of the nanoparticle.

Reading .xyz files

DAC.read_xyz — Method

read_xyz(filename::String)

Takes a path to a ".xyz" file and loads it as a Cluster type, which is returned.

Arguments

filename::String: Filename to load.

DAC.read_xyzs — Method

read_xyzs(filename::String, formula::Dict{String, Int64})

Reads in a .xyz file containing multiple molecules and returns a Vector of Cluster instances. All molecules must have the same formula, and this formula must also be provided as an argument.

Arguments

filename::String: Atomic coordinates.

formula::Dict{String, Int64}: Chemical formula of molecules.

Writing .xyz files

DAC.write_xyz — Method

write_xyz(filename::String, atoms::Cluster)

Takes a path (should end in ".xyz") to a file to write and saves a Cluster type to that path as an .xyz file.

Arguments

filename::String: Filename to save to.

atoms::Cluster: The nanoparticle to save.

DAC.write_xyz — Method

write_xyz(filename::String, positions::Matrix{Float64}, formula::Dict{String, Int64}, 
		cell::Float64)

Takes a path (should end in ".xyz") to a file to write and saves the given atomic coordinates, formula, and cell size (cubic) to that file.

Arguments

filename::String: Filename to save to.

positions::Matrix{Float64}: Atomic coordinates to save

formula::Dict{String, Int64}: Chemical formula.

cell::Float64: Length of the cubic unit cell.

DAC.write_xyz — Method

write_xyz(filename::String, atoms::Cluster, tags::Vector{Int64})

Takes a path (should end in ".xyz") to a file to write and saves a Cluster type to that path as a .xyz file. Also stores the tags of each atom.

Arguments

filename::String: Filename to save to.

atoms::Cluster: The nanoparticle to save.

tags::Vector{Int64}: the atom tags.

DAC.write_xyz — Method

write_xyz(filename::String, positions::Matrix{Float64}, formula::Dict{String, Int64}, 
		cell::Float64, tags::Vector{Int64})

Takes a path (should end in ".xyz") to a file to write and saves the given atomic coordinates, formula, and cell size (cubic) to that file.

Arguments

filename::String: Filename to save to.

positions::Matrix{Float64}: Atomic coordinates to save

formula::Dict{String, Int64}: Chemical formula.

cell::Float64: Length of the cubic unit cell.

tags::Vector{Int64}: the atom tags.

Generating a random Cluster

DAC.generateRandomSeed — Function

generateRandomSeed(formula::Dict{String, Int64}, boxLength::Number, 
		vacuumAdd::Number, returnCoordsOnly::Bool=false, 
		oherencyDistance::Float64=1.5)

Generates a seed with randomised atomic coordinates. Returns this either as a Cluster or simply the coordinates at Matrix{Float64}, debending on the value of returnCoordsOnly.

The coordinates are generated in a cubic cell with side lengths boxLength. Additional vacuum padding of lenghth vacuumAdd is added after atomic coordinates are generated. The final structure is also checked to make sure all atoms are connected, e.g. if one had two atom pair, with the members of each pair separated by 1 $\AA$ but the pairs themselves separated by 2 $\AA$, the nanoparticle would be conisdered incoherent if coherencyDistance was less than 2.

Arguments

formula::Dict{String, Int64}: Chemical formula.

boxLength::Number: Length of box to generate coordinates in.

vacuumAdd::Number: Amount of vacuum padding to add to the cell.

returnCoordsOnly::Bool: Will return a Cluster if false (default), or only the atomic coordinates if true.

coherencyDistance::Float64: Maximum distance any atom/group of atoms can be seperated from another group before the cluster is considered non-coherent.

Examples

To generate a random Au$_{55}$ Cluster:

generateRandomSeed(Dict("Au" => 55), 5.0, 10.0, false, 4.0)

To take an existing Au$_{55}$ Cluster called atoms and randomise is coordinates:

setPositions!(atoms, generateRandomSeed(Dict("Au" => 55), 5.0, 10.0, true, 4.0))