ClusterVector

When the BHA or DACA finish running, they will store all unique nanoparticle structures discovered to a file called clusterVector.jld2. To access these structures again, one will need to create a Julia script to read this file. The command below will load this back into Julia runtime.

julia> clusterVector = DAC.jldopen("clusterVector.jld2")["clusterVector"]

The ClusterCompressed type

DAC.ClusterCompressed — Type

ClusterCompressed

The ClusterCompressed type is for storing the structures of nanoparticles found during global optimisation. Only the positions, energy, and CNA profile of the structure is stored, along side an identifier.

Fields

positions: Atomic positions.

energy: Energy of the nanoparticle (rounded by the BasinHopper).

CNA: The CNA profile.

ID: ID of the nanoparticle.

ClusterVector and ClusterVectorWithML

DAC.ClusterVector — Type

ClusterVector

The ClusterVector is essentially a database of all nanoparticles found in the course of a global optimisation run. The BHA/DACA will store structures here in a data-race safe fashion (i.e. if multiple threads are performing walks in parallel, chaos is prevented).

Structures are stored in the vec field and are ordered from lowest to highest energy by the BHA/DACA. The BHA/DACA will not add duplicate structures, where two structures are considered duplicate if their CNA profiles are identical, and their energies are the same to a given decimal place (the dp is specified to the BasinHopper instance.) Only ounded energies are stored.

Note that structures are not stored as a Cluster type, but rather as ClusterCompressed.

Fields

vec::Vector{ClusterCompressed}: Vector of ClusterCompressed objects.

N::Threads.Atomic{Int64}: Number of clusters in the Vector (Atomic, datarace safe). To access the value of N itself (as opposed to the Atomic instance) use ClusterVector.N[].

lock::ReentrantLock: A lock to prevent data races by parallel threads.

DAC.ClusterVectorWithML — Type

ClusterVectorWithML

The ClusterVectorWithML is similar to ClusterVector but contains additional information regarding the Machine-Learning data used to classify the structure by DACA.

Fields

vec::Vector{ClusterCompressed}: Vector of ClusterCompressed objects.

eLim::Float64: The energy limit above which a cluster is not used for training.

MLData::Matrix{UInt8}: Data used for training, e.g. number of atoms belonging to Atom-64-Class

nMLData::Int64:

idsOfMLLabels::Vector{Vector{Int32}}:

idToIndex::Vector{Int32}:

N::Threads.Atomic{Int64}: Number of clusters in the Vector (atomic, datarace safe).

lock::ReentrantLock: A lock to prevent data races.