CNA Profiles and the SCM

The DAC package can calculate total and normal CNA profiles of nanoparticles. The following aliases are used as types to store these profiles.

DAC.CNAProfile — Type

const CNAProfile = Vector{Pair{Tuple{UInt8, UInt8, UInt8}, UInt16}}

Alias type for a total CNA profile, Vector{Pair{Tuple{UInt8, UInt8, UInt8}, UInt16}}

DAC.normalCNAProfile — Type

const normalCNAProfile = Vector{Dict{Tuple{UInt8, UInt8, UInt8}, UInt16}}

Alias type for Vector{Dict{Tuple{UInt8, UInt8, UInt8}, UInt16}}

Calculating a CNA profile

Note that all CNA calculation methods take in a coordinates argument (Matrix{Float64}) but also have wrapper functions that allow Cluster types to be used as arguments.

The following methods can be used to calculate CNA profiles.

DAC.getCNAProfile — Method

getCNAProfile(coordinates::Matrix{Float64}, rcut::Float64)

Calculates the total CNA profile for the cluster described by coordinates and returns it as a sorted Vector (sorted by the signature tuples as this slightly speeds computation for a Vector, which was used instead of a Dict to minimise storage space).

Arguments

coordinates::Matrix{Float64}: Atomic coordinates.

rcut::Float64: $r_\mathrm{cut}$ value used to define bond distance cut-off.

DAC.getNormalCNAProfile — Method

getNormalCNAProfile(coordinates::Matrix{Float64}, rcut::Float64)

Calculates the normal CNA profiles for the cluster described by coordinates and returns it as a Vector of Dict's, respectively. normalCNA[1] is the atom level CNA profile of the atom at coordinates[1, :]

Arguments

coordinates::Matrix{Float64}: Atomic coordinates.

rcut::Float64: $r_\mathrm{cut}$ value used to define bond distance cut-off

DAC.getTotalAndNormalCNAProfile — Method

getTotalAndNormalCNAProfile(coordinates::Matrix{Float64}, rcut::Float64)

Calculates the total and normal CNA profile for the cluster described by coordinates and returns both

Arguments

coordinates::Matrix{Float64}: Atomic coordinates.

rcut::Float64: $r_\mathrm{cut}$ value used to define bond distance cut-off.

Calculating SCM similarity.

DAC.getCNASimilarity — Method

getCNASimilarity(x::CNAProfile, y::CNAProfile)

Returns as a Float (0-1) the similarity between two total CNA profiles, x and y according to the Jaccard similarity index.