Metropolis Criteria

Metropolic Criteria types must implement the method getAcceptanceBoolean(MetC::EnergyMetC, oldCluster::Cluster, newCluster::Cluster) which returns a Tuple with the acceptance or rejection of the given step (i.e. true or false), and a String for debugging that denotes the chance to accept the hop (if any) and any other relavent information.

Energy Metropolis Criterion

DAC.EnergyMetC — Type

EnergyMetC

The Energy Metropolis Criterion. Compares the energy of an oldCluster and newCluster to accept or reject the hop with:

$A = \exp\left(\frac{E_\mathrm{old} - E_\mathrm{new}}{kT}\right)$

An instance of this Metropolic Criterion can be called with:

EnergyMetC(kT, io)

Fields

kT::Float64: reduced temperature

io::Tuple{IO, Channel}: IO channel for log.

HISTO Metropolis Criterion

DAC.HISTOMetC — Type

HISTOMetC

The HISTO Metropolis Criterion. Compares the energy of an oldCluster and newCluster in addition to the heights of the histogram bars each cluster belongs to in feature space. Here, the similarity to a reference structure is used as the feature space. This reference structure is selected in situ as the lowest energy structure found after a specifed number of hops (the waitTime).

$A = \exp\left[\frac{E_\mathrm{old} - E_\mathrm{new} + w\left(h_\mathrm{old} - h+\mathrm{new}\right)}{kT}\right]$

Fields

kT::Float64: reduced temperature

w::Float64: Wieght given to the memory contribution

delta::Float64: Histogram bar width in feature space.

resetPeriod::Float64: Redundant.

refCNA::CNAProfile: CNA profile used as a reference for the similarity histogram.

refID::Int64: ID of reference structure used for refCNA.

clusterVector::ClusterVector: Pointer to database of structures found by BHA.

waitTime::Int64: Number of hops to wait until setting the reference structure.

timeElapsed::Int64: Number of hops performed since the BHA started.

hist::Vector{Int64}: The histogram bar heights.

io::Tuple{IO, Channel}: IO channel for log.

DAC.HISTOMetC — Method

HISTOMetC(kT::Float64, w::Float64, delta::Float64, resetPeriod::Float64, clusterVector::ClusterVector, waitTime::Int64, io::Tuple{IO, Channel})

Wrapper function for HISTOMetC that sets some of the fields to initial empty states.

GMM Metropolis Criterion

@docs GMMMetC GMMMetC(gaussian::GMM, gaussianCluster::Int64, pca::PCA, mode::Symbol, useExplorationDataOnly::Bool, kT::Float64, io::Tuple{IO, Channel}) GMMMetC(gaussian::GMM, gaussianCluster::Int64, pca::PCA, mode::Symbol, useExplorationDataOnly::Bool, kT::Float64, classes::normalCNAProfile, io::Tuple{IO, Channel})