Displacement Operators

The following displacement operators are available for perturbing a Cluster.

DAC.cartesianDisplacement — Method

cartesianDisplacement(coords::Matrix{Float64}, dr::Float64)

Returns positions of atoms after moving each atom in each Cartesian direction by ±dr

Arguments

coords::Matrix{Float64}: Atomic coordinates (a wrapper function exists if a Cluster type is used instead).

dr::Float64 distance to move each atom in each direction.

DAC.geometricCentreDisplacement — Method

geometricCentreDisplacement(
	coords::Matrix{Float64}, alphaMin::Float64, 
	alphaMax::Float64, w::Float64
	)

Displaces atoms more the further away from the cluster centre they are. Didn't use this much and don't really remember how it works so good luck :/

DAC.perturbClusterSurface — Method

perturbClusterSurface(coords::Matrix{Float64}, nAtomsToMove::Int64, rCut::Float64)

Returns new postions of a cluster after moving nAtomsToMove atoms on the surface. nAtomsToMove can be a number (number of atoms) or float < 1 (% of total atoms). Only the least coordinated atoms are selected. If the number of lowest coordination atoms is less than nAtomsToMove then only this lower number of atoms are displaced (e.g. if nAtomsToMove == 2, and the lowest coordination number of an atom is 5, but only one such atom exists, only that atom will move). Atoms are moved to a random location on the cluster surface.

Arguments

coords::Matrix{Float64}: The atomic coordinates (a wrapper function exists if a Cluster type is used instead).

nAtomsToMove::Number: The number of atoms to move.

rCut::Float64: The bond distance cutoff definition.

Building Composite Operators

The BasinHopper contains a perturber field, this being a function called to perturb the structure of the Cluster for each attempted hop. Even if only a single displacement operator is being used, one must define a custom function for this. Below is an example of such a function which uses two different operators, selected at random. The function must return the new coordinates, and a String describing which operator was used.

dr = 1.0
nAtomsToMove = 1
rcut = 3.4765
function perturber(coords::Matrix{Float64})
    r = rand()
    if r < 0.75 # this is the likelihood of the CDO being used for perturbation
        return DC.cartesianDisplacement(coords, dr), "CDO perturbation"
    end
    
    return DC.perturbClusterSurface(coords, nAtomsToMove, rcut), "SAO perturbation" 
end