Reseeding

Reseeding is where the search restarts from a random starting structure periodically. There are currently two Reseeder types implemented. Any new types must flfull the following requirements.

An updateHopsToReseed! method must be implemented to track and update how many hops are left until a reseed is triggered.

A checkNewlyAcceptedStructure! method must be implemented so that, when a new structure is accepted, the Reseeder can check if this triggers any behaviour (e.g. if a new LES is found, this may reset how many hops are left until the next reseed occurs.)

A timeToReseed! function must be implemented. This function must return true or false depending on if it is time for a reseed to be performed.

A getReseedPeriod function must be implemented and should simply return the reseed period of the Reseeder.

The Reseeder must contain a field named getReseedStructure if timeToReseed! can return true. (If this sounds odd, note that the ReseedDisabled Reseeder will always return false on timeToReseed!). getReseedStructure is a Function passed to the Reseeder to generate a new random structure upon a reseed being triggered. Typically, getReseedStructure is set to generateRandomSeed, an inbuilt function to return the structure of a random seed. Additionally, an args field must be defined for the Reseeder containing all the arguments needed for the getReseedStructure function. args will then be unpacked using the ... operator for the getReseedStructure function.

Implemented Reseeder types

DAC.NewLESReseeder — Type

NewLESReseeder

Will trigger a reseed after reseedPeriod if no new lowest energy structure (since the last reseed) is found within that period. This period resets when a new LES is found.

Fields

reseedPeriod::Int64: The period after the latest LES discovery until a reseed is triggered.

hopsToReseed::Int64: The hops until the next reseed is triggered if no LES is found.

reseedEnergyToBeat::Float64: Tracks the latest LES energy.

getReseedStructure::Function: Function called if a reseed is triggered to generate the new seed.

args::Vector{Any}: Function arguments passed to getReseedStructure.

Example

reseedPeriod = 50
reseedEnergyToBeat = Inf # any structure will have a lower energy than +.v.e infty.
formula = Dict("Au" => 55)
boxLength = 5.0
vacuumAdd = 10.0
returnCoordsOnly = true
coherencyDistance = 4.0
getReseedStructure = generateRandomSeed
args = [formula, boxLength, vacuumAdd, returnCoordsOnly, coherencyDistance]
reseeder = NewLESReseeder(reseedPeriod, reseedPeriod, reseedEnergyToBeat, 
					getReseedStructure, args)

DAC.ReseedDisabled — Type

ReseedDisabled

Will never trigger a reseed.

Example

reseeder = ReseedDisabled()