Calculators

Two Calculator types are implemented in DAC, described below. New Calculator types can be defined, and need only implement two methods: calculateForces! and calculateEnergies!. The requirements are as follows.

calculateForces!

The method calculateForces!(atoms::Cluster, calc::Calculator) must set Cluster.forces to the force experienced by each atom in atoms. Cluster.validForces should also be set to true. This method should return nothing.

calculateEnergies!

The method calculateEnergies!(atoms::Cluster, calc::Calculator) must set Cluster.energies to the energy experienced by each atom in atoms. Additionall, atoms.energy should be set to the final sim of atoms.energies. Cluster.validEnergies should also be set to true. This method should return nothing.

Lennard-Jones

DAC.LJ — Type

LJ

Lennard-Jones Calculator.

Arguments

epsilon::Float64: epsilon parameter of the Lennard-Jones potential.

sigma::Float64: sigma parameter of the Lennard-Jones potential.

rc::Float64: Cut-off distance.

distance_vectors::Matrix{Float64}: Space in memory for force calculations.

r2::Vector{Float64}: Space in memory for force calculations.

c6::Vector{Float64}: Space in memory for force calculations.

c12::Vector{Float64}: Space in memory for force calculations.

DAC.LJ — Method

LJ(epsilon::Float64, sigma::Float64, rc::Float64, N::Int64)

Wrapper function for Lennard-Jones Calculator, that more easily reserves the space needed in memory based off the number of atoms, N. For reduced units, $\epsilon \ = \ \sigma \ = \ 1$

Arguments

epsilon::Float64: epsilon parameter of the Lennard-Jones potential.

sigma::Float64: sigma parameter of the Lennard-Jones potential.

rc::Float64: Cut-off distance.

N::Int64: Number of atoms in Atoms object. Used to reserve space in memory.

RGL (or Gupta)

DAC.RGL — Type

RGL

RGL Calculator.

Arguments

A::Float64: Attractive coefficient

p::Float64: Range of attractive coefficient

q::Float64: Range of repulsive coefficient

r0::Float64: Equilibrium bond distance

xi::Float64: $\xi$, repulsive coefficient

xi2::Float64: $\xi^2$ square of repulsive coefficient.

φ::Float64: Space in memory

DAC.RGL — Method

RGL(A::Float64, p::Float64, q::Float64, r0::Float64, xi::Float64, natoms::Int64)

Wrapper function for RGL Calculator, that more easily reserves the space needed in memory based off the number of atoms, natoms.

Arguments

A::Float64: Attractive coefficient

p::Float64: Range of attractive coefficient

q::Float64: Range of repulsive coefficient

r0::Float64: Equilibrium bond distance

xi::Float64: $\xi$, repulsive coefficient

natoms::Int64: Number of atoms.